Update .xprofile
This config is to if I'm using dwm (the which you can find a dedicated repo on my profile) use syspowermanager's options to run as daemon, monitor battery's charge and lid state and use slock as screenlocker to when the system is going to sleep. Otherwise, only use daemon and monitor options.