Updated Linux config file.
This config has some changes compared to the previous one, now it's more
adapted to my daily use. Some options which are required by some
applications I use were enabled.

This config file is for Gentoo Linux with systemd as its init system.
To use it on OpenRC just go to Gentoo wiki on the systemd guide,
specifically, and just remove the options listed there from this config
and you're ready to go.